Move AR adapter related helper methods under AR::TestCase
These helpers are called from both test class definition and test execution,
so we're defining them as both class and instance methods on AR::TestCase.
